swp/requirements.txt
Django==2.1.2
djangorestframework==3.8.2
mysqlclient==1.3.12
pytz==2018.5
